+++ src/signals/IncidentMap/components/FilterPanel/index.ts
@@ -1,0 +1,2 @@
+export { FilterPanel } from './FilterPanel'
+export type { Filter } from './FilterPanel'
