+++ src/signals/IncidentMap/components/index.ts
@@ -1,3 +1,3 @@
-export { FilterCategoryPanel } from './FilterCategoryPanel'
+export { FilterPanel } from './FilterPanel'
 export { IncidentLayer } from './IncidentLayer'
 export { IncidentMap } from './IncidentMap'
